+++ libraries/ui/src/main/java/androidx/media3/ui/PlayerView.java
@@ -887,8 +887,8 @@ public class PlayerView extends FrameLayout implements AdViewProvider {
   /**
    * Sets the {@link PlayerControlView.VisibilityListener}.
    *
-   * <p>Removes any listener set by {@link
-   * #setControllerVisibilityListener(PlayerControlView.VisibilityListener)}.
+   * <p>If {@code listener} is non-null then any listener set by {@link
+   * #setControllerVisibilityListener(PlayerControlView.VisibilityListener)} is removed.
    *
    * @param listener The listener to be notified about visibility changes, or null to remove the
    *     current listener.
@@ -896,14 +896,16 @@ public class PlayerView extends FrameLayout implements AdViewProvider {
   @SuppressWarnings("deprecation") // Clearing the legacy listener.
   public void setControllerVisibilityListener(@Nullable ControllerVisibilityListener listener) {
     this.controllerVisibilityListener = listener;
-    setControllerVisibilityListener((PlayerControlView.VisibilityListener) null);
+    if (listener != null) {
+      setControllerVisibilityListener((PlayerControlView.VisibilityListener) null);
+    }
   }
 
   /**
    * Sets the {@link PlayerControlView.VisibilityListener}.
    *
-   * <p>Removes any listener set by {@link
-   * #setControllerVisibilityListener(ControllerVisibilityListener)}.
+   * <p>If {@code listener} is non-null then any listener set by {@link
+   * #setControllerVisibilityListener(ControllerVisibilityListener)} is removed.
    *
    * @deprecated Use {@link #setControllerVisibilityListener(ControllerVisibilityListener)} instead.
    */
@@ -923,8 +925,8 @@ public class PlayerView extends FrameLayout implements AdViewProvider {
     this.legacyControllerVisibilityListener = listener;
     if (listener != null) {
       controller.addVisibilityListener(listener);
+      setControllerVisibilityListener((ControllerVisibilityListener) null);
     }
-    setControllerVisibilityListener((ControllerVisibilityListener) null);
   }
 
   /**